data-pipeline/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.3
python-dotenv==1.0.0
